Issue #738 replace .inc configuration files by .ttl (turtle) files.
The cache that was created in the Model instance, now was moved to the
GlobalConfig, but is still accessed from Model to add entries looked up.

There is no more need of a separate .ttl file for vocabularies, as it is
now part of the configuration file. Old files were removed.

In order to maintain backward compatibility, most methods of GlobalConfig
and of VocabularyConfig were maintained as-is. With internal changes to
re-route requests to the right RDF objects.

Tests were minor updates in tests, but only to replace .inc by .ttl or fix
coverage annotations.