Update index.html
new links added to the examples list to the right panel. The screen image was moved from right to the left panel.  One panel on the right is still empty.
on the left panel the button and the image were enclosed in <p> tags  to separate them from each other.
Example links to TIB and Rhineland-Palatinate were removed due to no response.